resize factor is shared util
